requirements.txt
flask
flask-cors
paho-mqtt
python-dotenv
